fixed unit tests randomly breaking
