+++ test/DataQI.EntityFrameworkCore.Test/Fixtures/DbFixture.cs
@@ -1,6 +1,7 @@
 using Microsoft.EntityFrameworkCore;
 using DataQI.EntityFrameworkCore.Repository.Support;
 using DataQI.EntityFrameworkCore.Test.Repository.Persons;
+using Microsoft.Data.Sqlite;
 
 namespace DataQI.EntityFrameworkCore.Test.Fixtures
 {
@@ -8,8 +9,7 @@ namespace DataQI.EntityFrameworkCore.Test.Fixtures
     {
         public DbFixture()
         {
-            Context = CreateTestContext();
-            TryCreateDabase();
+            Context = TestContext.NewInstance();
 
             // 1. Default
             // PersonRepository = new PersonRepository(Context);
@@ -17,27 +17,6 @@ namespace DataQI.EntityFrameworkCore.Test.Fixtures
             // 2. Provided
             var repositoryFactory = new EntityRepositoryFactory(Context);
             PersonRepository = repositoryFactory.GetRepository<IPersonRepository>();
-        }
-
-        private TestContext CreateTestContext()
-        {
-            var optionsBuilder = new DbContextOptionsBuilder();
-            optionsBuilder.UseSqlite("Data Source=TestDb.sqlite3");
-
-            var context = new TestContext(optionsBuilder.Options);
-            return context;
-        }
-
-        private bool TryCreateDabase()
-        {
-            try
-            {
-                return Context.Database.EnsureCreated();
-            }
-            catch
-            {
-                return false;
-            }
         }
 
         public TestContext Context { get; }
